Further decoupled userprofile functionality from RTPBackend app into userapi app
NOTE: any userprofile objects created before migration 0004_userprofile_user will result in a conflict.
Solution: manually delete these objects and create them after applying the abovementioned migration.
